requirements.txt
pyvirtualdisplay
mss
Pillow
opencv-python-headless
numpy
ultralytics